+++ fabric/src/main/java/net/mehvahdjukaar/polytone/fluid/fabric/FluidPropertiesManagerImpl.java
@@ -4,6 +4,7 @@ import net.fabricmc.fabric.api.client.render.fluid.v1.FluidRenderHandler;
 import net.fabricmc.fabric.api.client.render.fluid.v1.FluidRenderHandlerRegistry;
 import net.mehvahdjukaar.polytone.Polytone;
 import net.mehvahdjukaar.polytone.colormap.Colormap;
+import net.mehvahdjukaar.polytone.colormap.IColorGetter;
 import net.mehvahdjukaar.polytone.fluid.FluidPropertyModifier;
 import net.minecraft.client.color.block.BlockColor;
 import net.minecraft.core.registries.BuiltInRegistries;
@@ -16,7 +17,7 @@ public class FluidPropertiesManagerImpl {
         FluidRenderHandler handler = reg.get(fluid);
         if (!(handler instanceof PolytoneFluidRenderHandlerWrapper) && fluidMod.hasColormap()) {
             BlockColor c = fluidMod.getColormap();
-            if (c instanceof Colormap ccm) {
+            if (c instanceof IColorGetter ccm) {
                 c = Polytone.COLORMAPS.getOrCreateConcurrentColormap(ccm);
             }
             reg.register(fluid, new PolytoneFluidRenderHandlerWrapper(handler, c));