+++ changelog.md
@@ -1,1 +1,1 @@
-- made translucency work in gui
+- added biome compound composite colormap type